+++ InSituVisualization/Tagging/PerformanceTaggerProvider.cs
@@ -11,12 +11,11 @@
 
 using System;
 using System.ComponentModel.Composition;
-using InSituVisualization.Tagging;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Tagging;
 using Microsoft.VisualStudio.Utilities;
 
-namespace InSituVisualization
+namespace InSituVisualization.Tagging
 {
     [Export(typeof(ITaggerProvider))]
     [ContentType("text")]
